+++ ares-coupon/ares-coupon-permission-sdk/src/main/java/com/hank/ares/model/dto/CheckPermissionReqDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CheckPermissionReqDto {
-    private Long userId;
+    private Integer userId;
     private String uri;
     private String httpMethod;
 }
